LPS-54341 Move to kernel path to be consistent with EditorConstants.java class
