+++ modules/apps/blogs/blogs-portlet-editor-config/src/com/liferay/blogs/portlet/editor/config/BlogsPortletTextEditorConfigContributor.java
@@ -14,11 +14,11 @@
 
 package com.liferay.blogs.portlet.editor.config;
 
+import com.liferay.portal.kernel.editor.config.PortletEditorConfigContributor;
 import com.liferay.portal.kernel.json.JSONFactoryUtil;
 import com.liferay.portal.kernel.json.JSONObject;
 import com.liferay.portal.kernel.portlet.LiferayPortletResponse;
 import com.liferay.portal.theme.ThemeDisplay;
-import com.liferay.portlet.editor.config.PortletEditorConfigContributor;
 
 import java.util.Map;
