+++ portal-impl/src/com/liferay/portal/editor/config/PortletEditorConfigImpl.java
@@ -12,8 +12,10 @@
  * details.
  */
 
-package com.liferay.portlet.editor.config;
+package com.liferay.portal.editor.config;
 
+import com.liferay.portal.kernel.editor.config.PortletEditorConfig;
+import com.liferay.portal.kernel.editor.config.PortletEditorConfigContributor;
 import com.liferay.portal.kernel.json.JSONFactoryUtil;
 import com.liferay.portal.kernel.json.JSONObject;
 import com.liferay.portal.kernel.portlet.LiferayPortletResponse;
