+++ portal-service/src/com/liferay/portal/kernel/editor/config/PortletEditorConfigContributor.java
@@ -12,15 +12,27 @@
  * details.
  */
 
-package com.liferay.portlet.editor.config;
+package com.liferay.portal.kernel.editor.config;
+
+import com.liferay.portal.kernel.json.JSONObject;
+import com.liferay.portal.kernel.portlet.LiferayPortletResponse;
+import com.liferay.portal.theme.ThemeDisplay;
 
 import java.util.Map;
 
 /**
  * @author Sergio González
  */
-public interface PortletEditorConfig {
+public interface PortletEditorConfigContributor {
 
-	public Map<String, Object> getData();
+	public void populateConfigJSONObject(
+		JSONObject jsonObject, Map<String, Object> inputEditorTaglibAttributes,
+		ThemeDisplay themeDisplay,
+		LiferayPortletResponse liferayPortletResponse);
+
+	public void populateOptionsJSONObject(
+		JSONObject jsonObject, Map<String, Object> inputEditorTaglibAttributes,
+		ThemeDisplay themeDisplay,
+		LiferayPortletResponse liferayPortletResponse);
 
 }
